fix(cost-query-set): refactor costQuerySet request logic
