+++ apps/web/src/services/cost-explorer/CostExplorerContainer.vue
@@ -19,11 +19,20 @@
 </template>
 
 <script lang="ts">
+import { onUnmounted, watch } from 'vue';
+import { useRoute } from 'vue-router/composables';
+
+import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
+import { getCancellableFetcher } from '@cloudforet/core-lib/space-connector/cancallable-fetcher';
+
 import { useBreadcrumbs } from '@/common/composables/breadcrumbs';
+import ErrorHandler from '@/common/composables/error/errorHandler';
 import GeneralPageLayout from '@/common/modules/page-layouts/GeneralPageLayout.vue';
 import VerticalPageLayout from '@/common/modules/page-layouts/VerticalPageLayout.vue';
 
 import CostExplorerLNB from '@/services/cost-explorer/CostExplorerLNB.vue';
+import { useCostQuerySetStore } from '@/services/cost-explorer/store/cost-query-set-store';
+
 
 
 export default {
@@ -35,6 +44,49 @@ export default {
     },
     setup() {
         const { breadcrumbs } = useBreadcrumbs();
+        const costQuerySetStore = useCostQuerySetStore();
+        const costQuerySetState = costQuerySetStore.$state;
+        const route = useRoute();
+
+        onUnmounted(() => {
+            costQuerySetStore.$dispose();
+            costQuerySetStore.$reset();
+        });
+
+        watch(() => route.params, async (params) => {
+            // Case - Directly access Budget Page
+            if (!costQuerySetState.selectedDataSourceId) {
+                const fetcher = getCancellableFetcher(SpaceConnector.clientV2.costAnalysis.dataSource.list);
+                try {
+                    const { status, response } = await fetcher({
+                        query: {
+                            only: ['data_source_id'],
+                        },
+                    });
+                    if (status === 'succeed') {
+                        const dataSourceId = response.results[0].data_source_id;
+                        costQuerySetStore.$patch({
+                            selectedDataSourceId: dataSourceId,
+                        });
+                    }
+                } catch (e) {
+                    ErrorHandler.handleError(e);
+                }
+            }
+
+            /*
+            * Both parameters are set in the route. (beforeEnter navigation guard in route.ts)
+            * */
+            if (params.dataSourceId && params.costQuerySetId) {
+                costQuerySetStore.$patch({
+                    selectedDataSourceId: params.dataSourceId,
+                    selectedQuerySetId: params.costQuerySetId,
+                });
+            }
+            console.debug('[Cost Explorer]', costQuerySetState.selectedDataSourceId);
+            await costQuerySetStore.listCostQuerySets();
+        }, { immediate: true });
+
 
         return {
             breadcrumbs,
