+++ apps/web/src/services/cost-explorer/store/cost-query-set-store.ts
@@ -10,7 +10,7 @@ import ErrorHandler from '@/common/composables/error/errorHandler';
 import { managedCostQuerySets } from '@/services/cost-explorer/cost-analysis/config';
 import type { CostQuerySetModel } from '@/services/cost-explorer/type';
 
-const fetcher = getCancellableFetcher(SpaceConnector.client.costAnalysis.costQuerySet.list);
+const fetcher = getCancellableFetcher(SpaceConnector.clientV2.costAnalysis.costQuerySet.list);
 
 interface CostAnalysisLNBState {
     costQuerySetList: CostQuerySetModel[];
@@ -32,7 +32,6 @@ export const useCostQuerySetStore = defineStore('cost-query-set', {
     },
     actions: {
         async listCostQuerySets(): Promise<void> {
-            // TODO: apply v2
             try {
                 const { status, response } = await fetcher({
                     data_source_id: this.selectedDataSourceId,
